Rearrage files
